+++ volume-cartographer/core/include/vc/fiber_tracer/FiberTrace.hpp
@@ -20,7 +20,7 @@ struct FiberTraceConfig {
     double coneAngleDegrees = 25.0;
     double coneAngleStepDegrees = 5.0;
     int beamWidth = 8;
-    int beamLookaheadSteps = 1;
+    int beamLookaheadSteps = 2;
     double smoothnessWeight = 2.0;
     double smoothnessNormalWeight = 0.1;
     double smoothnessTangentWeight = 10.0;